Extract joystick utilities
