+++ Joystick/.settings/com.st.stm32cube.ide.mcu.sfrview.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+sfrviewstate={"fFavorites"\:{"fLists"\:{}},"fProperties"\:{"fNodeProperties"\:{}}}
